Working on data access layer. Refactoring.

+++ src/main/scala/models/Asset.scala
@@ -16,7 +16,7 @@ package models
   * @param decimals         scaling decimal places for record keeping.
   * @param display_decimals scaling decimal places for output display.
   */
-case class Asset(aclass: String, altname: String, decimals: Double, display_decimals: Double)
+case class Asset(aclass: String, altname: String, decimals: Int, display_decimals: Int)
